d1904-0922

+++ app/Controller/EbooksController.php
@@ -188,14 +188,16 @@ class EbooksController extends AppController
             return $this->redirect(array('action' => 'index'));
         }
         elseif ($this->Ebook->delete()) {
-            unlink(WWW_ROOT.'files/'.$data['Ebook']['user_id'].'/'.$data['Ebook']['file']);
-            if (pathinfo($data['Ebook']['file'], PATHINFO_EXTENSION)!= 'pdf'){
-                unlink(WWW_ROOT.'files/'.$data['Ebook']['user_id'].'/'.pathinfo($data['Ebook']['file'], PATHINFO_FILENAME).'.pdf');
+            if (file_exists(WWW_ROOT . 'files/' . $data['Ebook']['user_id'] . '/' . $data['Ebook']['file'])) {
+                unlink(WWW_ROOT . 'files/' . $data['Ebook']['user_id'] . '/' . $data['Ebook']['file']);
+                if (pathinfo($data['Ebook']['file'], PATHINFO_EXTENSION) != 'pdf') {
+                    unlink(WWW_ROOT . 'files/' . $data['Ebook']['user_id'] . '/' . pathinfo($data['Ebook']['file'], PATHINFO_FILENAME) . '.pdf');
+                }
+                unlink(WWW_ROOT . 'files/' . $data['Ebook']['user_id'] . '/pre_' . pathinfo($data['Ebook']['file'], PATHINFO_FILENAME) . '.pdf');
+                unlink(WWW_ROOT . 'files/' . $data['Ebook']['user_id'] . '/' . $data['Ebook']['picture']);
+                $this->Flash->success(__('Xóa thành công'));
+                return $this->redirect(array('action' => 'index'));
             }
-            unlink(WWW_ROOT.'files/'.$data['Ebook']['user_id'].'/pre_'.pathinfo($data['Ebook']['file'], PATHINFO_FILENAME).'.pdf');
-            unlink(WWW_ROOT.'files/'.$data['Ebook']['user_id'].'/'.$data['Ebook']['picture']);
-            $this->Flash->success(__('Xóa thành công'));
-            return $this->redirect(array('action' => 'index'));
         }
         $this->Flash->error(__('Xảy ra lỗi'));
         return $this->redirect(array('action' => 'index'));
